Separate parent modules for serial workflow and workflow engine
