+++ flowserv/controller/serial/worker/docker.py
@@ -15,8 +15,8 @@ from typing import Dict, Optional
 import logging
 import os
 
-from flowserv.controller.serial.result import ExecResult
-from flowserv.controller.serial.step import ContainerStep
+from flowserv.controller.serial.workflow.result import ExecResult
+from flowserv.controller.serial.workflow.step import ContainerStep
 from flowserv.controller.serial.worker.base import ContainerEngine
 
 import flowserv.util as util
@@ -61,7 +61,7 @@ class DockerWorker(ContainerEngine):
 
         Returns
         -------
-        flowserv.controller.serial.result.ExecResult
+        flowserv.controller.serial.workflow.result.ExecResult
         """
         # Keep output to STDOUT and STDERR for all executed commands in the
         # respective attributes of the returned execution result.
