+++ tests/controller/serial/worker/test_docker_worker.py
@@ -15,12 +15,12 @@ import os
 import pytest
 
 from flowserv.controller.serial.worker.docker import DockerWorker
-from flowserv.controller.serial.step import ContainerStep
+from flowserv.controller.serial.workflow.step import ContainerStep
 
 
 # Template directory
 DIR = os.path.dirname(os.path.realpath(__file__))
-RUN_DIR = os.path.join(DIR, '../../.files')
+RUN_DIR = os.path.join(DIR, '../../../.files')
 
 
 # -- Patching for error condition testing -------------------------------------
