+++ tests/controller/serial/worker/test_worker_config.py
