+++ tests/controller/serial/worker/test_worker_factory.py
