+++ tests/controller/serial/workflow/test_serial_workflow.py
@@ -10,8 +10,8 @@
 
 import pytest
 
-from flowserv.controller.serial.step import ContainerStep
-from flowserv.controller.serial.workflow import SerialWorkflow
+from flowserv.controller.serial.workflow.step import ContainerStep
+from flowserv.controller.serial.workflow.base import SerialWorkflow
 from flowserv.model.parameter.string import String
 
 import flowserv.error as err
